ai-tutor-backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.0

# Database
motor>=3.7.0
beanie==1.26.0

# Auth
python-jose[cryptography]==3.3.0
authlib==1.3.0
itsdangerous==2.2.0

# AI
anthropic==0.40.0

# HTTP client
httpx==0.27.0

# SSE streaming
sse-starlette==2.1.0

# Config
pydantic-settings==2.4.0

# Rate limiting
slowapi==0.1.9

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
